requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
jinja2
python-dotenv
flask
flask-sqlalchemy
pytest